Moved UI to a diffrent folder to seperate it from the game files.
